pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openpyxl>=3.1.5",
    "pillow>=12.0.0",
    "pycryptodome>=3.23.0",
    "python-dotenv>=1.2.1",
    "qrcode>=8.2",
    "werkzeug>=3.1.4",
]
